+++ lib/src/main/java/dev/doglog/internal/log_thread/entries/BooleanArrayQueuedLogEntry.java
@@ -8,7 +8,7 @@ public class BooleanArrayQueuedLogEntry extends BaseQueuedLogEntry {
   public final boolean[] value;
 
   public BooleanArrayQueuedLogEntry(String key, long timestamp, boolean[] value) {
-    super(key, timestamp);
+    super(EntryType.BOOLEAN_ARRAY, key, timestamp);
     this.value = value;
   }
 }