+++ lib/src/main/java/dev/doglog/internal/log_thread/entries/FloatArrayQueuedLogEntry.java
@@ -8,7 +8,7 @@ public class FloatArrayQueuedLogEntry extends BaseQueuedLogEntry {
   public final float[] value;
 
   public FloatArrayQueuedLogEntry(String key, long timestamp, float[] value) {
-    super(key, timestamp);
+    super(EntryType.FLOAT_ARRAY, key, timestamp);
     this.value = value;
   }
 }