+++ lib/src/main/java/dev/doglog/internal/log_thread/entries/FloatQueuedLogEntry.java
@@ -8,7 +8,7 @@ public class FloatQueuedLogEntry extends BaseQueuedLogEntry {
   public final float value;
 
   public FloatQueuedLogEntry(String key, long timestamp, float value) {
-    super(key, timestamp);
+    super(EntryType.FLOAT, key, timestamp);
     this.value = value;
   }
 }