+++ test-project/src/main/java/frc/robot/RobotContainer.java
@@ -4,6 +4,7 @@
 
 package frc.robot;
 
+import edu.wpi.first.wpilibj.PowerDistribution;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
 import dev.doglog.DogLog;
@@ -12,6 +13,7 @@ import dev.doglog.DogLogOptions;
 public class RobotContainer {
   public RobotContainer() {
     DogLog.setOptions(new DogLogOptions().withNtPublish(true));
+    DogLog.setPdh(new PowerDistribution());
 
     configureBindings();
   }