+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [2024.5.5](https://github.com/jonahsnider/doglog/compare/2024.5.4...2024.5.5) (2024-07-27)
+
+
+### Bug Fixes
+
+* stop automatically logging PDH to avoid CAN errors ([22340bf](https://github.com/jonahsnider/doglog/commit/22340bf802db8a4795a99c0f5420292a55151d3e))
+
+## [2024.5.4](https://github.com/jonahsnider/doglog/compare/2024.5.3...2024.5.4) (2024-07-27)
+
+
+### Performance Improvements
+
+* improve performance when processing queued log entries ([9c95541](https://github.com/jonahsnider/doglog/commit/9c9554180c39b127828e5b449a141bdb7a0ac445))
+
 ## [2024.5.3](https://github.com/jonahsnider/doglog/compare/2024.5.2...2024.5.3) (2024-07-22)